isis: extract LSP / CSNP / packet enums into lsp.rs
Move the LSP origination chain and packet-handler glue out of inst.rs
into a new lsp.rs:

  - target_block_name, target_locator_name (SR subscription helpers)
  - resolve_dis_ifindex (pseudonode lookup)
  - dis_generate (pseudonode LSP build)
  - lsp_generate (self-originated router LSP build, the biggest
    chunk — TLV assembly for area/proto/hostname/SR-MPLS/SRv6/MT/
    IS-reach/IP-reach)
  - lsp_emit (LSP → bytes + checksum patch)
  - csnp_generate (CSNP chunking from LSDB)
  - PacketMessage / Packet enums
  - lsp_flood (SRM-set trigger)

Consumers now pull these from super::lsp instead of super::inst:
ifsm.rs, network.rs, link.rs, packet.rs, lsdb.rs, and inst.rs itself.
mod.rs re-exports Packet/PacketMessage from lsp rather than inst.

The three SR-subscription / DIS helper unit tests move with their
functions; build_rib_nexthop / make_rib_entry tests stay in inst.rs
(they'll move with the RIB extraction in PR 4).

inst.rs drops from 2835 to 2030 lines.
